refactor(agent-task): 라우트 헬퍼 분리 — 600줄 CI 가드 준수
리뷰 반영(local 가드·pending 선리셋)으로 618줄이 되어 Gate 3 초과 —
loadOwnedTask/toPublicTask 를 agent-task.helpers.ts 로 분리(596줄).
