+++ apps/api/src/routes/agent-task.routes.ts
@@ -51,6 +51,7 @@ import {
 } from '../services/agent-task/upload-store';
 import { claimUploadsAsInputFiles, ChunkStoreError } from '../services/agent-task/chunk-store';
 import { resolveDefaultMaxTurns } from '../services/agent-task/task-inputs';
+import { loadOwnedTask, toPublicTask } from './agent-task.helpers';
 
 const logger = createLogger('AgentTaskRoutes');
 const router = Router();
@@ -59,29 +60,6 @@ const router = Router();
 router.use(requireAuth);
 
 type UserRole = 'admin' | 'user' | 'guest';
-
-/** 소유권 검증 후 작업 반환 — 없거나 권한 없으면 응답 종료하고 undefined 반환 */
-async function loadOwnedTask(req: Request, res: Response, taskId: string) {
-    const db = getUnifiedDatabase();
-    const task = await db.getAgentTask(taskId);
-    if (!task) {
-        res.status(404).json(notFound('작업을 찾을 수 없습니다.'));
-        return undefined;
-    }
-    assertResourceOwnerOrAdmin(String(task.user_id), String(req.user!.id), req.user!.role || 'user');
-    return task;
-}
-
-/** 응답용 변환: 큰 checkpoint/input_files/input_images 본문 제거 + resumable 플래그(중단된 작업에 체크포인트 존재).
- *  input_files 는 내용(content/data)을 뺀 메타(name/type/size)만 노출 — 목록/상세 응답 팽창 방지. */
-function toPublicTask(t: Record<string, unknown>) {
-    const { checkpoint, input_files, input_images, ...rest } = t;
-    void input_images; // dataURL 배열 — 응답에서 제외(팽창 방지)
-    const fileMetas = Array.isArray(input_files)
-        ? (input_files as AgentTaskInputFile[]).map((f) => ({ name: f?.name, type: f?.type, size: f?.size }))
-        : undefined;
-    return { ...rest, ...(fileMetas ? { input_files: fileMetas } : {}), resumable: !!checkpoint && t.status === 'failed' };
-}
 
 /**
  * POST /api/agent-tasks
